requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-environ>=0.11
Pillow>=10.0
